refactor(web): isolate sign-in state from chat
